+++ NEWS.md
@@ -2,6 +2,7 @@
  - Development version
  - `sim.default` now accepts the argument `R` to be a list (of list) of
  arguments.
+ - `subset.estimate` method
 
 # lava 1.8.0
  - New methods `estimate.mlm`, `IC.mlm`, `pars.mlm`, `estimate.array`,